refactor: reorganize upload modules
