--- components/upload/__init__.py
@@ -1,3 +1,0 @@
-from .ui_builder import UploadUIBuilder
-
-__all__ = ["UploadUIBuilder"]
