+++ services/interfaces.py
@@ -89,7 +89,7 @@ def get_upload_validator(
     c = _get_container(container)
     if c and c.has("upload_validator"):
         return c.get("upload_validator")
-    from services.upload.core.validator import ClientSideValidator
+    from services.upload.validator import ClientSideValidator
 
     return ClientSideValidator()
