+++ services/upload/__init__.py
@@ -12,7 +12,7 @@ from validation.security_validator import SecurityValidator
 
 from .ai import AISuggestionService, analyze_device_name_with_ai
 from .controllers.upload_controller import UnifiedUploadController as UploadController
-from .core.processor import UploadProcessingService
+from .processor import UploadProcessingService
 from .helpers import save_ai_training_data
 from .upload_core import UploadCore
 from .upload_types import ValidationResult, UploadResult
