+++ services/upload/core_api.py
@@ -5,7 +5,7 @@ from validation.file_validator import FileValidator
 from .file_processor_service import FileProcessor
 from .learning_coordinator import LearningCoordinator
 from .processor import UploadOrchestrator, UploadProcessingService
-from components.upload.ui_builder import UploadUIBuilder
+from components.ui_builder import UploadUIBuilder
 from .validator import ClientSideValidator
 
 __all__ = [
