+++ services/upload/orchestrator.py
@@ -8,10 +8,9 @@ import pandas as pd
 from core.protocols import FileProcessorProtocol
 from services.async_file_processor import AsyncFileProcessor
 from services.data_enhancer.mapping_utils import get_ai_column_suggestions
-from services.upload.core.file_processor_service import FileProcessor
-from services.upload.core.file_validator import FileValidator
-from services.upload.core.learning_coordinator import LearningCoordinator
-from components.upload.ui_builder import UploadUIBuilder
+from services.upload.file_processor_service import FileProcessor
+from services.upload.learning_coordinator import LearningCoordinator
+from components.ui_builder import UploadUIBuilder
 from services.upload.protocols import (
     DeviceLearningServiceProtocol,
     UploadProcessingServiceProtocol,
