+++ services/upload/service_registration.py
@@ -17,11 +17,11 @@ from services.interfaces import (
     UploadDataServiceProtocol,
 )
 from services.upload.controllers.upload_controller import UnifiedUploadController
-from services.upload.core.file_processor_service import FileProcessor
-from services.upload.core.learning_coordinator import LearningCoordinator
-from services.upload.core.processor import UploadProcessingService
-from components.upload.ui_builder import UploadUIBuilder
-from services.upload.core.validator import ClientSideValidator
+from services.upload.file_processor_service import FileProcessor
+from services.upload.learning_coordinator import LearningCoordinator
+from services.upload.processor import UploadProcessingService
+from components.ui_builder import UploadUIBuilder
+from services.upload.validator import ClientSideValidator
 from services.upload.protocols import (
     UploadControllerProtocol,
     UploadProcessingServiceProtocol,
