+++ services/upload/validators.py
@@ -1,5 +1,5 @@
 """Backward compatibility wrapper for ClientSideValidator."""
 
-from .core.validator import ClientSideValidator
+from .validator import ClientSideValidator
 
 __all__ = ["ClientSideValidator"]
